refactor: split actions.ts, migrate logger in monitor.ts and send.ts
- Split actions.ts (901 lines) into 5 files:
  - actions.ts (171 lines): main adapter with routing
  - actions-utils.ts (365 lines): shared helpers
  - actions-send.ts (28 lines): send action
  - actions-messages.ts (161 lines): read/edit/delete/react/pin/search
  - actions-admin.ts (383 lines): channel/user/org admin actions
- Migrated 14/15 core.log calls in monitor.ts to logger?.info/error
- Migrated all 5 core.log calls in send.ts to zulipLogger?.info
- Moved logger init outside try block for catch/finally access
- Removed unused formatZulipLog import from send.ts
- All 131 tests pass, typecheck clean, build succeeds
