+++ .pi-subagents/artifacts/2e0121a3_worker_0_meta.json
@@ -1,0 +1,127 @@
+{
+  "runId": "2e0121a3",
+  "agent": "worker",
+  "task": "In the project at /Volumes/external-hd/workspace/openclaw-zulip-bridge, I need to migrate `core.log` calls to the proper logger in `src/zulip/monitor.ts`.\n\nThe file is at src/zulip/monitor.ts (868 lines). There are 15 `core.log` calls that need to be replaced with `core.logging?.getChildLogger`.\n\nThe pattern to follow:\n- A `logger` variable already exists at line 81-82: `const logger = core.logging?.getChildLogger ? core.logging.getChildLogger({ module: \"zulip\" }) : null;`\n- A `logVerboseMessage` function already exists at line 84\n- Replace `core.log?.(formatZulipLog(\"...\", {...}))` with `logger?.info?.(\"...\", {...})` using the same data object\n- Replace `core.error?.(formatZulipLog(\"...\", {...}))` with `logger?.error?.(\"...\", {...})`\n\nExample:\n```typescript\n// OLD:\ncore.log?.(formatZulipLog(\"zulip monitor starting\", { accountId: opts.accountId }));\n// NEW:\nlogger?.info?.(\"zulip monitor starting\", { accountId: opts.accountId });\n```\n\nIMPORTANT: Do NOT change the `core.log` calls that are inside the `logInboundDrop` callback at line 375-382. That one uses `core.log` as a function parameter passed to `logInboundDrop`, not as a direct call.\n\nAlso, the `formatZulipLog` import is used in many places beyond just `core.log` calls (e.g., `core.error`, `logVerboseMessage`). Do NOT remove the import.\n\nRead the file first, find all 15 `core.log` calls, and replace them. Then verify with `npm run typecheck`.",
+  "exitCode": 143,
+  "usage": {
+    "input": 0,
+    "output": 0,
+    "cacheRead": 0,
+    "cacheWrite": 0,
+    "cost": 0,
+    "turns": 0
+  },
+  "model": "ollama/deepseek-v4-flash:cloud:high",
+  "attemptedModels": [
+    "ollama/deepseek-v4-flash:cloud:high"
+  ],
+  "modelAttempts": [
+    {
+      "model": "ollama/deepseek-v4-flash:cloud:high",
+      "success": false,
+      "exitCode": 143,
+      "usage": {
+        "input": 0,
+        "output": 0,
+        "cacheRead": 0,
+        "cacheWrite": 0,
+        "cost": 0,
+        "turns": 0
+      }
+    }
+  ],
+  "durationMs": 152016,
+  "toolCount": 0,
+  "launchContractDigest": "1352e832ef8a49133ad7a70b27734d96c3862f8ac02ff4cba8cf469424080664",
+  "acceptance": {
+    "status": "rejected",
+    "evidenceStatus": "rejected",
+    "explicit": false,
+    "effectiveAcceptance": {
+      "level": "checked",
+      "explicit": false,
+      "inferredReason": [
+        "risky write-capable run"
+      ],
+      "criteria": [
+        {
+          "id": "criterion-1",
+          "must": "Implement the requested change without widening scope",
+          "evidence": [
+            "changed-files",
+            "tests-added",
+            "commands-run",
+            "residual-risks",
+            "no-staged-files"
+          ],
+          "severity": "required"
+        },
+        {
+          "id": "criterion-2",
+          "must": "Return evidence sufficient for an independent acceptance review",
+          "evidence": [
+            "changed-files",
+            "tests-added",
+            "commands-run",
+            "residual-risks",
+            "no-staged-files"
+          ],
+          "severity": "required"
+        }
+      ],
+      "evidence": [
+        "changed-files",
+        "tests-added",
+        "commands-run",
+        "residual-risks",
+        "no-staged-files"
+      ],
+      "verify": [],
+      "review": {
+        "agent": "reviewer",
+        "required": true
+      },
+      "stopRules": []
+    },
+    "inferredReason": [
+      "risky write-capable run"
+    ],
+    "criteria": [
+      {
+        "id": "criterion-1",
+        "must": "Implement the requested change without widening scope",
+        "evidence": [
+          "changed-files",
+          "tests-added",
+          "commands-run",
+          "residual-risks",
+          "no-staged-files"
+        ],
+        "severity": "required"
+      },
+      {
+        "id": "criterion-2",
+        "must": "Return evidence sufficient for an independent acceptance review",
+        "evidence": [
+          "changed-files",
+          "tests-added",
+          "commands-run",
+          "residual-risks",
+          "no-staged-files"
+        ],
+        "severity": "required"
+      }
+    ],
+    "runtimeChecks": [
+      {
+        "id": "attestation",
+        "status": "failed",
+        "message": "Structured acceptance report not found."
+      }
+    ],
+    "verifyRuns": [],
+    "childReportParseError": "Structured acceptance report not found."
+  },
+  "transcriptPath": "/Volumes/external-hd/workspace/openclaw-zulip-bridge/.pi-subagents/artifacts/2e0121a3_worker_0_transcript.jsonl",
+  "timestamp": 1785603752411
+}
