+++ .pi-subagents/artifacts/2e0121a3_worker_0_transcript.jsonl
@@ -1,0 +1,1 @@
+{"version":1,"recordType":"message","source":"foreground","runId":"2e0121a3","agent":"worker","childIndex":0,"cwd":"/Volumes/external-hd/workspace/openclaw-zulip-bridge","ts":1785603600376,"timestamp":"2026-08-01T17:00:00.376Z","sourceEventType":"initial_prompt","role":"user","text":"In the project at /Volumes/external-hd/workspace/openclaw-zulip-bridge, I need to migrate `core.log` calls to the proper logger in `src/zulip/monitor.ts`.\n\nThe file is at src/zulip/monitor.ts (868 lines). There are 15 `core.log` calls that need to be replaced with `core.logging?.getChildLogger`.\n\nThe pattern to follow:\n- A `logger` variable already exists at line 81-82: `const logger = core.logging?.getChildLogger ? core.logging.getChildLogger({ module: \"zulip\" }) : null;`\n- A `logVerboseMessage` function already exists at line 84\n- Replace `core.log?.(formatZulipLog(\"...\", {...}))` with `logger?.info?.(\"...\", {...})` using the same data object\n- Replace `core.error?.(formatZulipLog(\"...\", {...}))` with `logger?.error?.(\"...\", {...})`\n\nExample:\n```typescript\n// OLD:\ncore.log?.(formatZulipLog(\"zulip monitor starting\", { accountId: opts.accountId }));\n// NEW:\nlogger?.info?.(\"zulip monitor starting\", { accountId: opts.accountId });\n```\n\nIMPORTANT: Do NOT change the `core.log` calls that are inside the `logInboundDrop` callback at line 375-382. That one uses `core.log` as a function parameter passed to `logInboundDrop`, not as a direct call.\n\nAlso, the `formatZulipLog` import is used in many places beyond just `core.log` calls (e.g., `core.error`, `logVerboseMessage`). Do NOT remove the import.\n\nRead the file first, find all 15 `core.log` calls, and replace them. Then verify with `npm run typecheck`.\n\n## Acceptance Contract\nAcceptance level: checked\nCompletion is not accepted from prose alone. End with a structured acceptance report.\n\nCriteria:\n- criterion-1: Implement the requested change without widening scope\n- criterion-2: Return evidence sufficient for an independent acceptance review\n\nRequired evidence: changed-files, tests-added, commands-run, residual-risks, no-staged-files\n\nReview gate: required by reviewer.\n\nFinish with a fenced JSON block tagged `acceptance-report` in this shape:\nUse empty arrays when no items apply; array fields contain strings unless object entries are shown.\n`criteriaSatisfied[].status` must be exactly one of: satisfied, not-satisfied, not-applicable.\n`commandsRun[].result` must be exactly one of: passed, failed, not-run.\n`manualNotes` and `notes` are optional strings; an empty string means no note and does not satisfy `manual-notes` evidence.\n```acceptance-report\n{\n  \"criteriaSatisfied\": [\n    {\n      \"id\": \"criterion-1\",\n      \"status\": \"satisfied\",\n      \"evidence\": \"specific proof\"\n    },\n    {\n      \"id\": \"criterion-2\",\n      \"status\": \"satisfied\",\n      \"evidence\": \"specific proof\"\n    }\n  ],\n  \"changedFiles\": [\n    \"src/file.ts\"\n  ],\n  \"testsAddedOrUpdated\": [\n    \"test/file.test.ts\"\n  ],\n  \"commandsRun\": [\n    {\n      \"command\": \"command\",\n      \"result\": \"passed\",\n      \"summary\": \"short result\"\n    }\n  ],\n  \"validationOutput\": [\n    \"validation output or concise summary\"\n  ],\n  \"residualRisks\": [\n    \"none\"\n  ],\n  \"noStagedFiles\": true,\n  \"diffSummary\": \"short description of the diff\",\n  \"reviewFindings\": [\n    \"blocker: file.ts:12 - issue found, or no blockers\"\n  ],\n  \"manualNotes\": \"anything else the parent should know\"\n}\n```","message":{"role":"user","content":[{"type":"text","text":"In the project at /Volumes/external-hd/workspace/openclaw-zulip-bridge, I need to migrate `core.log` calls to the proper logger in `src/zulip/monitor.ts`.\n\nThe file is at src/zulip/monitor.ts (868 lines). There are 15 `core.log` calls that need to be replaced with `core.logging?.getChildLogger`.\n\nThe pattern to follow:\n- A `logger` variable already exists at line 81-82: `const logger = core.logging?.getChildLogger ? core.logging.getChildLogger({ module: \"zulip\" }) : null;`\n- A `logVerboseMessage` function already exists at line 84\n- Replace `core.log?.(formatZulipLog(\"...\", {...}))` with `logger?.info?.(\"...\", {...})` using the same data object\n- Replace `core.error?.(formatZulipLog(\"...\", {...}))` with `logger?.error?.(\"...\", {...})`\n\nExample:\n```typescript\n// OLD:\ncore.log?.(formatZulipLog(\"zulip monitor starting\", { accountId: opts.accountId }));\n// NEW:\nlogger?.info?.(\"zulip monitor starting\", { accountId: opts.accountId });\n```\n\nIMPORTANT: Do NOT change the `core.log` calls that are inside the `logInboundDrop` callback at line 375-382. That one uses `core.log` as a function parameter passed to `logInboundDrop`, not as a direct call.\n\nAlso, the `formatZulipLog` import is used in many places beyond just `core.log` calls (e.g., `core.error`, `logVerboseMessage`). Do NOT remove the import.\n\nRead the file first, find all 15 `core.log` calls, and replace them. Then verify with `npm run typecheck`.\n\n## Acceptance Contract\nAcceptance level: checked\nCompletion is not accepted from prose alone. End with a structured acceptance report.\n\nCriteria:\n- criterion-1: Implement the requested change without widening scope\n- criterion-2: Return evidence sufficient for an independent acceptance review\n\nRequired evidence: changed-files, tests-added, commands-run, residual-risks, no-staged-files\n\nReview gate: required by reviewer.\n\nFinish with a fenced JSON block tagged `acceptance-report` in this shape:\nUse empty arrays when no items apply; array fields contain strings unless object entries are shown.\n`criteriaSatisfied[].status` must be exactly one of: satisfied, not-satisfied, not-applicable.\n`commandsRun[].result` must be exactly one of: passed, failed, not-run.\n`manualNotes` and `notes` are optional strings; an empty string means no note and does not satisfy `manual-notes` evidence.\n```acceptance-report\n{\n  \"criteriaSatisfied\": [\n    {\n      \"id\": \"criterion-1\",\n      \"status\": \"satisfied\",\n      \"evidence\": \"specific proof\"\n    },\n    {\n      \"id\": \"criterion-2\",\n      \"status\": \"satisfied\",\n      \"evidence\": \"specific proof\"\n    }\n  ],\n  \"changedFiles\": [\n    \"src/file.ts\"\n  ],\n  \"testsAddedOrUpdated\": [\n    \"test/file.test.ts\"\n  ],\n  \"commandsRun\": [\n    {\n      \"command\": \"command\",\n      \"result\": \"passed\",\n      \"summary\": \"short result\"\n    }\n  ],\n  \"validationOutput\": [\n    \"validation output or concise summary\"\n  ],\n  \"residualRisks\": [\n    \"none\"\n  ],\n  \"noStagedFiles\": true,\n  \"diffSummary\": \"short description of the diff\",\n  \"reviewFindings\": [\n    \"blocker: file.ts:12 - issue found, or no blockers\"\n  ],\n  \"manualNotes\": \"anything else the parent should know\"\n}\n```"}]}}
