+++ .pi-subagents/artifacts/3e24dca6_worker_0_output.md
@@ -1,0 +1,1 @@
+All validations pass. Here is the structured acceptance report.
